Renamed the notifcation folder (not the test one tho)
