+++ app/src/androidTest/java/io/github/polysmee/calendar/CalendarActivityPublicAppointmentsFragmentTest.java
@@ -23,7 +23,7 @@ import io.github.polysmee.calendar.calendarActivityFragments.CalendarActivityPub
 import io.github.polysmee.database.DatabaseFactory;
 import io.github.polysmee.login.AuthenticationFactory;
 import io.github.polysmee.login.MainUserSingleton;
-import io.github.polysmee.znotification.AppointmentReminderNotificationSetupListener;
+import io.github.polysmee.notification.AppointmentReminderNotificationSetupListener;
 
 import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed;
 import static com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn;
