+++ app/src/androidTest/java/io/github/polysmee/calendar/CalendarActivityTest.java
@@ -25,7 +25,7 @@ import io.github.polysmee.database.DatabaseFactory;
 import io.github.polysmee.invites.InvitesManagementActivity;
 import io.github.polysmee.login.AuthenticationFactory;
 import io.github.polysmee.login.MainUserSingleton;
-import io.github.polysmee.znotification.AppointmentReminderNotificationSetupListener;
+import io.github.polysmee.notification.AppointmentReminderNotificationSetupListener;
 
 import io.github.polysmee.room.RoomActivity;
