+++ app/src/androidTest/java/io/github/polysmee/roomActivityTests/RoomActivityNotParticipantTest.java
@@ -20,7 +20,7 @@ import io.github.polysmee.R;
 import io.github.polysmee.database.DatabaseFactory;
 import io.github.polysmee.login.AuthenticationFactory;
 import io.github.polysmee.login.MainUserSingleton;
-import io.github.polysmee.znotification.AppointmentReminderNotificationSetupListener;
+import io.github.polysmee.notification.AppointmentReminderNotificationSetupListener;
 import io.github.polysmee.room.RoomActivity;
 
 import static androidx.test.espresso.action.ViewActions.click;
