+++ app/src/androidTest/java/io/github/polysmee/roomActivityTests/RoomActivityParticipantsFragmentNotOwnerTest.java
@@ -16,7 +16,7 @@ import org.junit.runners.JUnit4;
 import io.github.polysmee.database.DatabaseFactory;
 import io.github.polysmee.login.AuthenticationFactory;
 import io.github.polysmee.login.MainUserSingleton;
-import io.github.polysmee.znotification.AppointmentReminderNotificationSetupListener;
+import io.github.polysmee.notification.AppointmentReminderNotificationSetupListener;
 import io.github.polysmee.room.fragments.RoomActivityParticipantsFragment;
 
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
