+++ app/src/androidTest/java/io/github/polysmee/roomActivityTests/RoomActivityParticipantsFragmentTest.java
@@ -20,7 +20,7 @@ import io.github.polysmee.database.DatabaseFactory;
 import io.github.polysmee.database.databaselisteners.BooleanChildListener;
 import io.github.polysmee.login.AuthenticationFactory;
 import io.github.polysmee.login.MainUserSingleton;
-import io.github.polysmee.znotification.AppointmentReminderNotificationSetupListener;
+import io.github.polysmee.notification.AppointmentReminderNotificationSetupListener;
 import io.github.polysmee.room.fragments.RoomActivityParticipantsFragment;
 
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
