+++ app/src/androidTest/java/io/github/polysmee/roomActivityTests/RoomActivityTest.java
@@ -24,7 +24,7 @@ import io.github.polysmee.appointments.AppointmentActivity;
 import io.github.polysmee.database.DatabaseFactory;
 import io.github.polysmee.login.AuthenticationFactory;
 import io.github.polysmee.login.MainUserSingleton;
-import io.github.polysmee.znotification.AppointmentReminderNotificationSetupListener;
+import io.github.polysmee.notification.AppointmentReminderNotificationSetupListener;
 import io.github.polysmee.room.RoomActivity;
 
 import static androidx.test.espresso.intent.matcher.IntentMatchers.hasExtra;
