+++ app/src/androidTest/java/io/github/polysmee/roomActivityTests/RoomActivityVideoFragmentTest.java
@@ -28,7 +28,7 @@ import io.github.polysmee.database.DatabaseFactory;
 import io.github.polysmee.database.databaselisteners.BooleanChildListener;
 import io.github.polysmee.login.AuthenticationFactory;
 import io.github.polysmee.login.MainUserSingleton;
-import io.github.polysmee.znotification.AppointmentReminderNotificationSetupListener;
+import io.github.polysmee.notification.AppointmentReminderNotificationSetupListener;
 import io.github.polysmee.room.RoomActivity;
 import io.github.polysmee.room.fragments.RoomActivityVideoFragment;
