+++ app/src/main/AndroidManifest.xml
@@ -40,7 +40,7 @@
 
         <activity android:name=".photo.editing.PictureEditActivity"/>
 
-        <receiver android:name=".znotification.AppointmentReminderNotificationPublisher" />
+        <receiver android:name=".notification.AppointmentReminderNotificationPublisher" />
 
         <activity android:name=".login.LoginActivity" />
         <activity android:name=".room.RoomActivity" />
