+++ app/src/main/java/io/github/polysmee/calendar/CalendarActivity.java
@@ -21,7 +21,7 @@ import io.github.polysmee.R;
 
 import io.github.polysmee.invites.InvitesManagementActivity;
 
-import io.github.polysmee.znotification.AppointmentReminderNotificationSetupListener;
+import io.github.polysmee.notification.AppointmentReminderNotificationSetupListener;
 import io.github.polysmee.settings.SettingsActivity;
 
 public class CalendarActivity extends AppCompatActivity{
