+++ app/src/main/java/io/github/polysmee/notification/AppointmentReminderNotificationPublisher.java
@@ -1,4 +1,4 @@
-package io.github.polysmee.znotification;
+package io.github.polysmee.notification;
 
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
