+++ app/src/main/java/io/github/polysmee/notification/AppointmentReminderNotificationSetupListener.java
@@ -1,4 +1,4 @@
-package io.github.polysmee.znotification;
+package io.github.polysmee.notification;
 
 import android.app.AlarmManager;
 import android.app.PendingIntent;
